requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
python-dotenv==1.0.1
anthropic==0.40.0
python-pptx==1.0.2
